+++ lib/are_search/searcher/query_builder/complex_field_query_builder.rb
@@ -17,9 +17,7 @@ module AreSearch
                     :query_string,
                     :fields,
                     :query_type,
-                    :mlt_instance,
-                    :mlt_index_target,
-                    :mlt_params,
+                    :mlt,
                 ].freeze
             end